backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.10.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
httpx>=0.26.0
numpy>=1.26.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-multipart>=0.0.9